% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAUClr.R
\name{calcAUClr}
\alias{calcAUClr}
\title{calcAUClr}
\usage{
calcAUClr(data, group, cores = NULL, X = NULL)
}
\arguments{
\item{data}{abundance matrix or data frame (rows are samples, columns are variables (taxa))}

\item{group}{a vector with the sample groups}

\item{cores}{a number of cores fo paralelization, if \code{cores=NULL}, \code{parallel::detectCores()-1} will be used}

\item{X}{a \eqn{n\times p} matrix of p covariates observed in each sample}
}
\value{
\code{res} the upper triangular of AUC between OTUS
}
\description{
parallel AUC for data
}
\examples{
data(HIV)
x_HIVImp = zCompositions::cmultRepl(x_HIV, method="GBM",
output="p-counts",suppress.print=TRUE,z.warning=0.99)
Xnp <- model.matrix(y_HIV~MSM_HIV)
AUC <- calcAUClr(data = x_HIVImp, group = y_HIV, cores=2, X=Xnp)
AUC[1:10,1:10]
}
