% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_fun.R
\name{detect_knee_piecewise}
\alias{detect_knee_piecewise}
\title{Approximate variance calculation for the separability index \(S_k\)}
\usage{
detect_knee_piecewise(y, min_k = 5, enforce_shape = TRUE)
}
\arguments{
\item{y}{Value fo function}

\item{min_k}{Integer. Maximum number of OTUs \eqn{k} for which to calculate the variance.
Defaults to \code{ncol(V_matrix)}.}

\item{enforce_shape}{description}
}
\value{
A numeric vector of length \code{max_k}, where the entry at position \eqn{k}
        contains the approximate variance of the separability index \(S_k\).
        The value for \eqn{k = 1} is \code{NA} as the index is undefined there.
}
\description{
Computes an approximate variance of the separability index \(S_k\) defined over
subsets of OTUs, accounting for a constant correlation \code{rho} between
pairwise AUC estimates.
}
\details{
This function is intended for internal use only.
}
\keyword{internal}
