% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRRelev.R
\name{LRRelev}
\alias{LRRelev}
\title{LRRelev}
\usage{
LRRelev(
  data,
  sample,
  group,
  taxa,
  otus,
  threshold = 2,
  cores = NULL,
  X = NULL,
  conf.level = 0.95,
  method = c("hanley", "delong", "bootstrap"),
  rho = 0.1,
  boot.n = 500
)
}
\arguments{
\item{data}{A data frame or matrix with counts: OTUs in columns, samples in rows.}

\item{sample}{A vector with the name or ID of each sample.}

\item{group}{A vector indicating the group of each sample.}

\item{taxa}{A vector with the taxonomic classification for each OTU.}

\item{otus}{A vector with the name of each OTU.}

\item{threshold}{Minimum total count for an OTU to be included. OTUs below this are excluded.}

\item{cores}{Number of cores for parallelization. If \code{NULL}, uses \code{parallel::detectCores()-1}.}

\item{X}{An optional \eqn{n \times p} matrix of covariates observed in each sample.}

\item{conf.level}{The confidence level for intervals, not in percent (e.g., 0.95 for 95\% CI).}

\item{method}{The method to use: \code{"hanley"}, \code{"delong"}, or \code{"bootstrap"}.}

\item{rho}{Mean correlation assumed between pairs of AUCs.}

\item{boot.n}{The number of bootstrap replicates. Default is 500.}
}
\value{
A list with:
\describe{
  \item{\code{data1Imp}}{Data with zeros imputed using GBM replacement.}
  \item{\code{OTUS}}{A data frame with OTUs and their association index.}
  \item{\code{Misery}}{A vector of OTUs with total counts less than the threshold.}
  \item{\code{uniqueOTUS}}{A vector of OTUs present in only one sample.}
  \item{\code{AUCs}}{The matrix of pairwise AUCs for OTUs.}
  \item{\code{OTUSRelev}}{A vector with the most relevant OTUs based on the index.}
}
}
\description{
Compute parallel log-ratio separability indices for microbiome data.
}
\examples{
# Example with HIV dataset:
data(HIV)
Xnp <- model.matrix(y_HIV ~ MSM_HIV)
output1 <- LRRelev(data = x_HIV, sample = rownames(x_HIV), group = y_HIV,
  taxa = colnames(x_HIV), otus = colnames(x_HIV), cores = 2,
   X = Xnp, method = "hanley")

}
