% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAUClr.R
\name{calcAUClr}
\alias{calcAUClr}
\title{calcAUClr}
\usage{
calcAUClr(data, group)
}
\arguments{
\item{data}{abundance matrix or data frame (rows are samples, columns are variables (taxa))}

\item{group}{a vector with the sample groups}
}
\value{
\code{res} the upper triangular of AUC between OTUS
}
\description{
parallel AUC for data
}
\examples{
data(HIV)
AUC <- calcAUClr(x_HIV, y_HIV)
AUC$`association log-ratio with y`
}
