% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAUClr.R
\name{calcAUClr}
\alias{calcAUClr}
\title{calcAUClr}
\usage{
calcAUClr(data, group, cores = NULL, binary = TRUE)
}
\arguments{
\item{data}{abundance matrix or data frame (rows are samples, columns are variables (taxa))}

\item{group}{a vector with the sample groups}

\item{cores}{a number of cores fo paralelization, if \code{cores=NULL}, \code{parallel::detectCores()-1} will be used}

\item{binary}{\code{logical} TRUE if group are binary, FALSE if group are multinomial}
}
\value{
\code{res} the upper triangular of AUC between OTUS
}
\description{
parallel AUC for data
}
\examples{
data(HIV)
AUC <- calcAUClr(x_HIV, y_HIV, cores=2)
AUC[1:10,1:10]
}
