% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAUClr.R
\name{calcAUClr}
\alias{calcAUClr}
\title{calcAUClr}
\usage{
calcAUClr(
  data,
  group,
  cores = NULL,
  X = NULL,
  conf.level = 0.95,
  method = c("hanley", "delong", "bootstrap"),
  rho = 0.1
)
}
\arguments{
\item{data}{Abundance matrix or data frame (rows are samples, columns are taxa/variables).}

\item{group}{A vector indicating the group of each sample.}

\item{cores}{Number of cores for parallelization. If \code{NULL}, uses \code{parallel::detectCores()-1}.}

\item{X}{An optional \eqn{n \times p} matrix of covariates for each sample.}

\item{conf.level}{The confidence level for intervals, not in percent (e.g., 0.95 for 95\% CI).}

\item{method}{The method to use: \code{"hanley"}, \code{"delong"}, or \code{"bootstrap"}.}

\item{rho}{Mean correlation assumed between pairs of AUCs.}
}
\value{
A list with:
\describe{
  \item{\code{AUC}}{The upper triangular matrix of pairwise AUCs between OTUs.}
  \item{\code{VAR}}{The upper triangular matrix of the variances of each AUC.}
}
}
\description{
Compute pairwise AUCs for all OTUs in compositional microbiome data.
}
\examples{
data(HIV)
x_HIVImp <- zCompositions::cmultRepl(x_HIV, method = "GBM",
  output = "p-counts", suppress.print = TRUE, z.warning = 0.99)
Xnp <- model.matrix(y_HIV ~ MSM_HIV)
AUC <- calcAUClr(data = x_HIVImp, group = y_HIV, cores = 2, X = Xnp,
method = "hanley")
AUC$AUC[1:10, 1:10]

}
