% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAUClr.R
\name{calcAUClr}
\alias{calcAUClr}
\title{calcAUClr}
\usage{
calcAUClr(
  data,
  group,
  cores = NULL,
  X = NULL,
  conf.level = 0.95,
  method = c("hanley", "delong", "bootstrap"),
  rho = 0.1
)
}
\arguments{
\item{data}{abundance matrix or data frame (rows are samples, columns are variables (taxa))}

\item{group}{a vector with the sample groups}

\item{cores}{a number of cores fo paralelization, if \code{cores=NULL}, \code{parallel::detectCores()-1} will be used}

\item{X}{a \eqn{n\times p} matrix of p covariates observed in each sample}

\item{conf.level}{the width of the confidence interval as [0,1], never in percent. Default: 0.95, resulting in a 95% CI.}

\item{method}{the method to use: \code{hanley} , \code{delong} or \code{bootstrap}.}

\item{rho}{Mean of correlation between pairs of AUC.}
}
\value{
\code{AUC} the upper triangular of AUC between OTUS

\code{VAR} the upper triangular of the variance of each AUC between OTUS
}
\description{
parallel AUC for data
}
\examples{
data(HIV)
x_HIVImp = zCompositions::cmultRepl(x_HIV, method="GBM",
output="p-counts",suppress.print=TRUE,z.warning=0.99)
Xnp <- model.matrix(y_HIV~MSM_HIV)
AUC <- calcAUClr(data = x_HIVImp, group = y_HIV, cores=2, X=Xnp)
AUC[1:10,1:10]
}
