% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRRelev.R
\name{LRRelev}
\alias{LRRelev}
\title{LRRelev}
\usage{
LRRelev(
  data,
  sample,
  group,
  taxa,
  otus,
  threshold = 2,
  cores = NULL,
  X = NULL,
  conf.level = 0.95,
  method = c("hanley", "delong", "bootstrap"),
  rho = 0.1,
  boot.n = 500
)
}
\arguments{
\item{data}{data with counts with OTUs in the columns and samples in rows}

\item{sample}{name of each sample}

\item{group}{a vector with the sample groups}

\item{taxa}{a vector with the taxon classification of each OTU}

\item{otus}{a vector with the name of each OTUS}

\item{threshold}{minimum count for the OTU entered for comparison}

\item{cores}{a number of cores fo paralelization, if \code{cores=NULL}, \code{parallel::detectCores()-1} will be used}

\item{X}{a \eqn{n\times p} matrix of p covariates observed in each sample}

\item{conf.level}{the width of the confidence interval as [0,1], never in percent. Default: 0.95, resulting in a 95% CI.}

\item{method}{the method to use: \code{hanley} , \code{delong} or \code{bootstrap}.}

\item{rho}{Mean of correlation between pairs of AUC.}

\item{boot.n}{the number of bootstrap replicates. Default: 500.}
}
\value{
\code{data1Imp} data with imputed zeros

\code{OTUS} a dataframe with otus and their association index Cesc calculate

\code{Misery} a list of OTUS with counts less than the threshold

\code{uniqueOTUS} a list of OTUS with counts only in one sample

\code{AUCs} The matrix of AUCs by OTUS

\code{OTUSRelev} The list of relevant otus
}
\description{
parallel logarithmic ratios
}
\examples{
data(HIV)
Xnp <- model.matrix(y_HIV~MSM_HIV)
output1 <- LRRelev(data=x_HIV, sample = rownames(x_HIV), group = y_HIV,
 taxa = colnames(x_HIV),otus = colnames(x_HIV), cores=2, X=Xnp)


}
