% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRRelev.R
\name{LRRelev}
\alias{LRRelev}
\title{LRRelev}
\usage{
LRRelev(
  data,
  sample,
  group,
  taxa,
  otus,
  binary = TRUE,
  threshold = 2,
  cores = NULL
)
}
\arguments{
\item{data}{data with counts with OTUs in the columns and samples in rows}

\item{sample}{name of each sample}

\item{group}{a vector with the sample groups}

\item{taxa}{a vector with the taxon classification of each OTU}

\item{otus}{a vector with the name of each OTUS}

\item{binary}{\code{logical} TRUE if group are binary, FALSE if group are multinomial}

\item{threshold}{minimum count for the OTU entered for comparison}

\item{cores}{a number of cores fo paralelization, if \code{cores=NULL}, \code{parallel::detectCores()-1} will be used}
}
\value{
\code{data1Imp} data with imputed zeros

\code{OTUS} a dataframe with otus and their association index Cesc calculate

\code{Misery} a list of OTUS with counts less than the threshold

\code{uniqueOTUS} a list of OTUS with counts only in one sample

\code{AUCs} The matrix of AUCs by OTUS
}
\description{
parallel logarithmic ratios
}
\examples{
data(HIV)
output1 <- LRRelev(data=x_HIV, sample = rownames(x_HIV), group = y_HIV,
 taxa = colnames(x_HIV),otus = colnames(x_HIV),binary=TRUE , cores=2)


}
