% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowlogratios.R
\name{rowlogratios}
\alias{rowlogratios}
\title{rowlogratios}
\usage{
rowlogratios(
  data,
  group,
  col,
  X = NULL,
  conf.level = 0.95,
  method = c("hanley", "delong", "bootstrap"),
  rho = 0.1
)
}
\arguments{
\item{data}{data with counts with OTUs in the columns and samples in rows}

\item{group}{a vector with the sample groups}

\item{col}{the number of column for calculate the logratios}

\item{X}{a \eqn{n\times p} matrix of p covariates observed in each sample}

\item{conf.level}{the width of the confidence interval as [0,1], never in percent. Default: 0.95, resulting in a 95% CI}

\item{method}{the method to use: 'hanley', 'delong' or 'bootstrap'}

\item{rho}{Mean of correlation between pairs of AUC}
}
\value{
\code{AUC} A vector of size \eqn{1\times m} with the AUC between the OTU's attainments in the given column

\code{VAR} A vector of size \eqn{1\times m} with the variances of AUC between the OTU's attainments in the given column
}
\description{
parallel logarithmic ratios
}
\examples{
data(HIV)
x_HIVImp <- zCompositions::cmultRepl(x_HIV, method="GBM",
output="p-counts",suppress.print=TRUE,z.warning=0.99)
Xnp <- model.matrix(y_HIV~MSM_HIV)
AUC <- rowlogratios(data= x_HIVImp, col= 2, group=y_HIV, X =Xnp)
AUC[1:10]

}
