% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowlogratios.R
\name{rowlogratios}
\alias{rowlogratios}
\title{rowlogratios}
\usage{
rowlogratios(
  data,
  group,
  col,
  X = NULL,
  conf.level = 0.95,
  method = c("hanley", "delong", "bootstrap"),
  rho = 0.1
)
}
\arguments{
\item{data}{A data frame or matrix with counts; OTUs in columns and samples in rows.}

\item{group}{A vector with the sample groups.}

\item{col}{The index (integer) of the column for which the logratios are computed.}

\item{X}{An optional \eqn{n \times p} matrix of covariates for each sample.}

\item{conf.level}{The confidence level for intervals, not in percent (e.g., 0.95 for 95\% CI).}

\item{method}{The method to use: \code{"hanley"}, \code{"delong"} or \code{"bootstrap"}.}

\item{rho}{Mean correlation assumed between pairs of AUCs.}
}
\value{
A list with:
\describe{
  \item{\code{AUC}}{A vector of size \eqn{1 \times m} with the AUCs comparing the selected OTU to all others.}
  \item{\code{VAR}}{A vector of size \eqn{1 \times m} with the variances of those AUCs.}
}
}
\description{
Compute pairwise logarithmic ratios for a selected OTU (column) in parallel.
}
\examples{
data(HIV)
x_HIVImp <- zCompositions::cmultRepl(x_HIV, method = "GBM",
  output = "p-counts", suppress.print = TRUE, z.warning = 0.99)
Xnp <- model.matrix(y_HIV ~ MSM_HIV)
AUC <- rowlogratios(data = x_HIVImp, col = 2, group = y_HIV,
 X = Xnp, method = "hanley")
AUC$AUC[1:10]

}
