% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowlogratios.R
\name{rowlogratios}
\alias{rowlogratios}
\title{rowlogratios}
\usage{
rowlogratios(data, col, group)
}
\arguments{
\item{data}{data with counts with OTUs in the columns and samples in rows}

\item{col}{the number of column for calculate the logratios}

\item{group}{a vector with the sample groups}
}
\value{
\code{matlrAUC} A vector of size \eqn{1\times n} with the OTU's attainments in the given column
}
\description{
parallel logarithmic ratios
}
\examples{
data(HIV)
AUC <- rowlogratios(x_HIV, 2, y_HIV)
AUC[1:10]

}
