% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowlogratios.R
\name{rowlogratios}
\alias{rowlogratios}
\title{rowlogratios}
\usage{
rowlogratios(data, col, group, X = NULL)
}
\arguments{
\item{data}{data with counts with OTUs in the columns and samples in rows}

\item{col}{the number of column for calculate the logratios}

\item{group}{a vector with the sample groups}

\item{X}{a \eqn{n\times p} matrix of p covariates observed in each sample}
}
\value{
\code{matlrAUC} A vector of size \eqn{1\times n} with the OTU's attainments in the given column
}
\description{
parallel logarithmic ratios
}
\examples{
data(HIV)
x_HIVImp <- zCompositions::cmultRepl(x_HIV, method="GBM",
output="p-counts",suppress.print=TRUE,z.warning=0.99)
Xnp <- model.matrix(y_HIV~MSM_HIV)
AUC <- rowlogratios(data= x_HIVImp, col= 2, group=y_HIV, X =Xnp)
AUC[1:10]

}
